requirements.txt
cement==3.0.4
jinja2
pyyaml
colorlog
netifaces==0.10.9
tabulate
print